data_code/requirements.txt
Keras==2.2.4
Keras-Applications==1.0.6
numpy==1.15.4
Pillow==5.4.1
pytest==4.6.5
tensorflow==1.12.0